fix/timezone-dependent-on-utc-plugin-comment
Moved comment `dependent on utc plugin` to the line with timezone because the timezone plugin is dependent on the utc plugin, not the utc plugin is dependent on the utc plugin.

It took me some time to figure out what I was doing wrong and why I'm getting `Cannot set property '$timezone' of undefined` error. After some searching, I found on gitter that the timezone plugin is dependent on the utc plugin, so your documentation needs a minor fix.